refactor(spend): dedupe Base rail helpers and refresh rail docs
Extract shared treasury→embedded recipient logic for conversion and Base USDC
rails; centralize embedded EVM and USDC amount guards in base-usdc-rail; align
SpendPaymentRail and session context JSDoc with current behavior; rename a
test for stable wording.
